requirements.txt
pymongo==3.9.0
requests==2.22.0
slackclient==2.3.1
